+++ includes/modules/trait-charitable-hooks.php
@@ -41,6 +41,10 @@ trait Charitable_Hooks {
 			return;
 		}
 
+		if ( wp_is_post_revision( $post_id ) || wp_is_post_autosave( $post_id ) ) {
+			return;
+		}
+
 		if ( 'campaign' !== get_post_type( $post_id ) ) {
 			return;
 		}