+++ includes/modules/trait-memberpress-hooks.php
@@ -37,6 +37,10 @@ trait MemberPress_Hooks {
 			return;
 		}
 
+		if ( wp_is_post_revision( $post_id ) || wp_is_post_autosave( $post_id ) ) {
+			return;
+		}
+
 		if ( 'memberpressproduct' !== get_post_type( $post_id ) ) {
 			return;
 		}